Fix a bug in hexedit.vim.
Closing the hex window is supposed to restore some settings to the
original buffer. This was done with a :buffer! command followed by
setting local options. But, if the current window was not the hex
window, the wrong buffer is swapped out by :buffer!. Apparently this
doesn't lose changes, but it does replace the current buffer
unexpectedly.