Restore window settings in more cases.
This was primarily to fix the case where the original buffer was closed
before the diff buffer, causing it to retain its diff settings if it was
ever re-displayed. However, it has the pleasant side-effect of also
fixing the bug listed under "Known Bugs", by restoring all windows
containing an instance of the buffer.

Unfortunately there's a new Known Bug, but it's more of a Vim feature.