Add resolve() in same-file check.
The same-file check wasn't working when symlinks were involved. Apparently
a:filename would have the symlinks resolved and expand("%") would not, so the
paths would be different even though the file was the same, and saving would
give an error (because it would look like you were attempting to overwrite an
existing, different file).

I'm not sure if something in Vim changed that caused this to break or if it's
been broken all along.