First pass adding XTerm bracketed paste mode.
This is not quite complete, lacking proper restoring of the options if 'term'
changes.

This requires moving the FixTerm call to the main script body, out of the
VimEnter autocmd, which is too late in the startup process. I don't see any
problem with this, but I don't recall why it was in the autocmd in the first
place (it may have simply been because the TermChanged autocmd was already being
used, so it was easy to add another event).