Don't use setmatches.
The getmatches/setmatches API has been broken, apparently intentionally. The
results returned from getmatches are no longer acceptable to setmatches if any
match was created with the new matchaddpos function. This results in errors in
this plugin when other matches are present (for example, if the cursor is on a
paren being highlighted by MatchPairs when jumping to the window).