Make matchpairs empty in hex buffers.
Since the hex buffer uses a different kind of pair matching (matching the hex
representation in the middle region with the ascii representation in the right
side region), it can conflict with the matchparen plugin. This results in an
error if the ascii region includes "{}" for example. The error is seen any time
the cursor moves onto a character that matchparen would try to highlight.