Fix E788 errors in fmtwarn.
This seems to be a problem in Vim 8, not seen in Vim 7.4. To test, start Vim
with no file, enter the command

:autocmd FileType * wincmd w

Followed by

:copen

Either the quickfix window will open with the cursor moved back into the
original buffer (as expected) or you'll get an E788 error. If you get the error,
then this fix is necessary to avoid the same error in fmtwarn.